+++ tests/parallel/conftest.py
@@ -1,18 +1,20 @@
+import joblib
 import pytest
 
-from pydvl.parallel.config import ParallelConfig
+from pydvl.parallel import JoblibParallelBackend, RayParallelBackend
 
 from ..conftest import num_workers
 
 
 @pytest.fixture(scope="module", params=["joblib", "ray-local", "ray-external"])
-def parallel_config(request):
+def parallel_backend(request):
     if request.param == "joblib":
-        yield ParallelConfig(backend="joblib", n_cpus_local=num_workers())
+        with joblib.parallel_config(backend="loky", n_jobs=num_workers()):
+            yield JoblibParallelBackend()
     elif request.param == "ray-local":
         ray = pytest.importorskip("ray", reason="Ray not installed.")
         ray.init(num_cpus=num_workers())
-        yield ParallelConfig(backend="ray")
+        yield RayParallelBackend()
         ray.shutdown()
     elif request.param == "ray-external":
         ray = pytest.importorskip("ray", reason="Ray not installed.")
@@ -24,6 +26,6 @@ def parallel_config(request):
             initialize_head=True, head_node_args={"num_cpus": num_workers()}
         )
         ray.init(cluster.address)
-        yield ParallelConfig(backend="ray", address=cluster.address)
+        yield RayParallelBackend()
         ray.shutdown()
         cluster.shutdown()